+++ src/Completions/Controllers/HtmlCompletionController.cs
@@ -114,10 +114,11 @@ namespace TailwindCSSIntellisense.Completions.Controllers
                     {
                         case VSConstants.VSStd2KCmdID.TYPECHAR:
                             var character = GetTypeChar(pvaIn);
-                            if (CharsAfterSignificantPoint(classText) <= 1 || character == ':' || character == '/')
+                            if (_currentSession == null || CharsAfterSignificantPoint(classText) <= 1 || character == ':' || character == '/')
                             {
                                 _currentSession?.Dismiss();
                                 StartSession(true);
+                                Filter();
                             }
                             else if (_currentSession != null)
                             {
@@ -130,10 +131,11 @@ namespace TailwindCSSIntellisense.Completions.Controllers
                             StartSession(true);
                             break;
                         case VSConstants.VSStd2KCmdID.BACKSPACE:
-                            if (CharsAfterSignificantPoint(classText) <= 2 || classText.EndsWith("/"))
+                            if (_currentSession == null || CharsAfterSignificantPoint(classText) <= 2 || classText.EndsWith("/"))
                             {
                                 _currentSession?.Dismiss();
                                 StartSession(true);
+                                Filter();
                             }
                             else if (_currentSession != null)
                             {
@@ -278,6 +280,7 @@ namespace TailwindCSSIntellisense.Completions.Controllers
             }
             else if (shouldStartNew)
             {
+                DismissOtherSessions();
                 _currentSession = Broker.CreateCompletionSession(TextView, snapshot.CreateTrackingPoint(caret, PointTrackingMode.Positive), true);
                 _currentSession.Dismissed += (sender, args) => _currentSession = null;
                 _currentSession.Start();